Verify sblim-sfcb successfully starts post upgrade
cmpi-provider-register script attempts to determine if sblim-sfcb
service is installed by detecting a link to the service start
symlink. Unfortunatelly, the name of the service was renamed in
SP1 resulting in cmpi-provider-register not restarting sfcb. Fixing
the naming issue resulted in the cmpi-provider-register from
working completely, which is a regression of 874811

These tests attempt to detect such a regression in the future.

See bnc#1041885 and bnc#874811