Switch to x11 console as precondition for x11 tests
In extra_tests_on_gnome we have tests for x11 which assume that we are
in x11 console, which is not always the case. For newly introduced
firefox_java tests we've added this step to x11regressions_setup test.

x11regressions_setup will also become lastgood, which is running x11 and
hence tests will not fail due to switching to lastgood with text
terminal.