Fix installation sequence for leap 42.3 maintenance
Recently we have created new tests for Leap 42.3 for maintenance and
updates. Existing tests are failing as expect addon_products step which
is no longer there after setting up online repos. For maintenance flavor
this step is exluded for leap 42.3. As we change naming convention, we
may need to adjust version check in future. For 42.2 addon_products step
is required.

See poo#16996