Investigate YaST2 failures by looking at y2log
Looking into the file y2log for any failure in yast related modules could
provide a very good first guess what the problem might be without needing to
transcribe screenshot content or waiting for a user to download the archive
file, unpack it, read the right file, scroll to the end, learn the format and
find a potentialy symptom. A 'record_info' window is added to job details
based on content in the file 'badlist' or a unique string being present in
y2log.

Verification run: http://lord.arch/tests/6705

The modules in the verification run are expected to fail, they show how the
error is investigated.