Remove variable expansion from tftp autoyast test
In aytests framework variables expansion is implemented, which doesn't
exist in openQA. But setup is static and we precisely know ip address of
supportserver. All sources are copied one to one from aytests repo.
We can get rid of support server here and upload tftp script simply by
using worker as we do for autoinst profiles. As of now, we make it this
way.