Remove variable expansion from sle12+support server
For autoyast test we don't use full functionality of aytests framework,
and variables are not get expanded. Hence as a temporary solution we
create profile copy with hardcoded url to the script which is available
on support server. Both were copied from original aytests repo:
https://github.com/yast/aytests-tests/tree/master/aytests

This is temporary solution, we should get rid of two copies of same
files.