LTP: Increase shutdown timeout to 1800 (30 mins)
The poweroff job in systemd has a timeout of 30 minutes on Tumbleweed. This
may seem excessive, but on the other hand I have never seen shutdown fail on
the LTP tests except due to a service being slow to exit. This results in a
cascade of failures, which we badly want to avoid.