Fix ftp test for autoyast regression
This test was never working. There is part which will get fixed in
bsc#1047232, another part is to perform upload as anonymous user. For
that proper permissions have to be set. For example, write operation has
to be forbidden on anon_root directory, but is allowed on
subdirectories. Which makes it impossible to upload to root ftp
directory as anonymous user. Additionally, we introduce changes to
creation mask of files uploaded by anonym in order to be able to
download it afterwords.

Test verified manually, after setting "listen=NO" in vsftpd.conf file.
After fix is there, test should work smoothly.