Start firewall and apache before testing ports
In the test we verify open ports in firewall rules, whereas we don't
have some of them added to iptables
before we start the firewall service.
This part is also adjusted according to the http.sh script located in
aytests. It cannot be reused as of now, as fails to verify ports. Hence
we have our local copy and back merge this script to aytests.