Replace hotkey usages for some screens in kde live
Initial problem was that some Next buttons got new shortcuts, in the end
we have identified that they are different in 2 runs of same
distribution running on different workers. If user goes back, next
button may also change it's shortcut.
So far it happens only on timezone settings page and network settings
page which is kde live specific.