Add new PXE boot specific tag and unload needles
As we have a lot of needles, we use ENV- tags to unregister irrelevant
needles in order to improve mathing performance. While fixing autoyast
tests we got another grub menu needle, which is specific to pxe_menu
boot. Here is tiny improvement to unload not used needles for other
tests.