Remove good_buttons from tests
Initial idea of the test case was to check release notes button. After a
while needles were created to match tag even though there was no release
button and test became useless. Test was also supposed to check screen
after welcome screen, but during execution it matches welcome screen
still. On top of it, for cryptlvm test on leap
42.2 we have expected pop-up which prevents needle from matching.
As this test case is useless, we remove it and add release notes button
to the 'install-overview' needles, which will be done separately both
for opensuse and sle.