Log available space on hd before installing apache
During one of the executions it was not possible to install apache due
to lack of space on the hard drive. HDD setting is set to 20GB, and
apache takes 15MB only which is not really expected. Even though it
doesn't happen often. On top of that mariadb was successfully installed
in next test case, even though it requires 150MB for installation.
Here we log disk usage in order to identify the root cause of the issue
if it reoccurs.