Readd StartXinted option to ftp autoyast profile
After discussion with autoyast team in commit
2fff9863946e96b78bbdb9b533faef6d450812b9 we have removed this option, as
it was not really used in the yast code. However, it is required
according ot the comment #3 in bsc#1047232. So we introduce it back.
Documentation for this option is already merged and will be available to
the customers, as was missing previously which was the reason to remove
it.