Add extra logging before reboot with gnome
In bsc#100059 we have an issue that logout dialog does not appear after
pressing ctrl-alt-delete keys. This issue is not easy to reproduce, and
now it mainly occurs on power and aarch. We were requested to provide
extra logs. I've tried to use shared ppc workers, but was not able to
reach reboot_gnome test as it fails before. After discussion we've
decided to introduce these steps in production and revert them as soon
as we have logs.
After increasing logging level for services, we also restart xdm and
then need to login again.