Add tests for libvirt-lxc and system container image
libvirt-lxc can be used to run containers. This test runs a system
container image as a libvirt lxc container. This has the double
advantage to test both the image and the libvirt lxc driver.

This test is to be run by maintenance to test SLES 12 SP3+ system
container images, but could also be used for validating the
libvirt-lxc containers feature on both SLES and openSUSE.