Add softfail for dns autoyast test
One of the reverse zones is not configured properly with given autoyast
profile which is a bug. Unfortunately, there is no chance that it gets
fixed in sp3, but should work in sle15. As we don't want to forget about
this issue, we don't exclude this check, but introduce soft-failure.