Fix profile user generation to make it consistent
In many profiles new user was added manually, and we just define
mandatory fields for them. Nevertheless for users which are already in
the profile we have more properties defined, including uid. The problem
is that autoyast is not aware of uid and simply tries to create users.
Uids are given sequentially starting from 1000. Meaning that if for one
user is generated automatically and next one has uid 1000 defined it
will lead to a conflict. Documentation is updated to warn users about
this behavior and this can be improved in future by doing some
verification before using profile.

See bsc#1047646.