Replace softfail with info for  partition warning
When we have introduced this soft-fail I was not aware that in some
profiles we have warning timeout configured explicitly, meaning we have
timeout on reduced partition screen. However, test will soft-fail with
bug reference that timeout is not shown. This commit improves behavior
and soft-fails only if there is no timeout, which we check with
wait_screen_change function call.
After discussion was decided simply to remove softfailure with
information window.