Make ftp tests more stable
Initial issue is that ftp tests are unstable on Leap. With introduced
changes we want to make it more stable. This is achieved by performing
checks that screen has changed after pressing certain key, etc. Problem
appears only on Leap, for SLE and TW tests are stable. The main issue
is that it's hard to say when ftp wizard has finished it's work.
Previously we used serial output to wait for, whereas 60 seconds timeout
isn't always enough. So main change here to fix tests is increase of
timeout, but to assert that console appeared.

See poo#20204.