Improve yast control center test stability on leap
Main improvement here is to use new set of needles which checks that
button persists in dialog and can be pressed. Previously, mainl header
of the window was checked, which in case of slow system can fail test as
key combination was pressed too early. Another major change was to
increase timeout in places where test fails during the execution on
production.