Improve yast gui proxy test stability
Yast ui tests are extremely unstable on leap. Whereas there are some
modules which are fragile due to the test flow. Proxy is one of them. On
of the issues here is that in previous implementation we hit buttons,
but do not verify result of our actions. And in case of bad performance
it happens that some actions were not performed as expeted. As the
result, test fails. Another issue is that it's also not clear why test
failed and what exactly went wrong.

See poo#20564.