Remove 'check_gcc48_on_sdk_in_aarch64' with low risk and low value
When I implemented the test case I was already not really convinced by the
value of it. poo#20098 shows that we have problems with that test module
during migration but we do not care so much about where a package actually
comes from. Also we confirmed the module is in the right repo for SLE 12 SP2
and stayed at the correct location during SLE 12 SP3. As the development for
SLE 12 SP3 is over now and for upcoming versions the version as well as the
package repo would change anyway we should simply remove this test module.

Related progress issue: https://progress.opensuse.org/issues/20098