Stabilize yast2_ui tests on openSUSE
Now I've changed approach to introducing changes to multiple tests in
one run according to my findings on the runs. Process reminds a bit
plugging holes in a sinking boat with fingers. On top of that chances
to spot new potential issue are higher when tests are executed as a
suite and not separately.

Please, be aware that one of the main points is to make test suite
reliable even though we compromise some extra checks.

See poo#19922 and all blocking tasks.