Add test modules to switch to a Plasma Wayland session
If the WAYLAND variable is enabled, the X11 tests run in a wayland session.
This requires two new test modules:

plasma5_force_96dpi forces the font DPI to 96 in Plasma sessions to keep
the same scale as X. This is necessary to match X11 needles in the wayland
session and does not have any impact outside of that.

start_wayland_plasma5 then installs plasma5-session-wayland (if necessary)
and applies a workaround for a crash in Mesa by setting the GALLIUM_DRIVER
environment variable to softpipe in the Plasma Wayland session.
It then logs out of the current session, selects the "Plasma 5 (Wayland)"
session in sddm, logs in again and waits for the desktop to appear.
This is necessary as both Live media and default installations have
autologin enabled by default.