Enable stress tests on s390x
As part of poo#18936 we plan to run stress tests on s390x. Whereas, to
boot to the image which is created in chained task some additional steps
have to performed on s390x architecture. Common parts of the code were
moved to bootloader_setup module to store them in one place. zkvm image
path variable is also stored there and is exported, not to have it
defined in multiple places.