Move or sync profiles and scripts with aytest repo
As part of improvement of autoyast tests, we want to keep profiles and
verification scripts in single place. As same profiles are used for
integration tests, we move them to aytests repository.
Whereas, as some files were forked, we need to sync them, to be on
the  same side. After merging these changes, settings of relevant jobs
have to be changed.
For remaining profiles we give them better names and get rid of support
server, if not required.