Remove version from yast profiles names
Autoyast profiles in autoyast_sle12sp2 folder were adjusted for sle12sp3
and name is irrelevant. Version 12 is not removed for these, as for sle
15 they may require adjustments as sle 15 have different architecture
and old profiles are not expected to work with sle 15 as of now. It may
be the case, that they will be automatically converted, which is not
trivial mapping. Hence leave version 12 in directory and some profiles
names.
Test suites have to be adjusted accordingly.