Accept not latest java version installed
When we test java on mozilla, occasionally, we have not latest version
installed. It also is not yet in official openSUSE repos, hence we
cannot simply update java. But as far as aim of the test is to verify
that java version is there, we can ignore this warning, as long as java
version is detected.

See poo#20632