Adjust the x11 VT number when switching to wayland in start_wayland_plasma5
For a Wayland session, sddm cannot reuse the greeter VT (tty7) as it is in
use by the X server, so it does the VT allocation itself. It queries the
kernel for the first free one and uses it. tty1 and tty2 are allocated during
consoletest_setup, so it is very likely to use tty3.