Fix last snapshot id parsing
Initially it looked like problem with typing, which we see a lot on
production. Whereas, in the test code we slice script output to get rid
of new line symbol in the end. Which is not there always. And using
script_output doesn't need this magic, and is used in other part of code
in same test.

See poo#21796