LTP: Update package list for SLE15 and remove WE softfail
Add curl because it is not installed on LeanOS by default. Move sysstat to
optional packages because it is missing on SLE15. Also move tpm-tools to
optional packages because the tests require more hardware support anyway.

Remove ntfsprogs soft fail, just install it if we can (it is currently
available in the SLE15 IBS repo).