Update bug refs in soft failures for SLE15
The initial bug is resolved as invalid and is expected behavior. So, we
should change it to real root cause of this workaround that kernel dev
packages are not available in Development-Tools modules repo. Due to
existing conflicts curl is also not installed, soft failure will help us
to track the state of it and resolve bug once we have curl installed.