Add workaround not to use tty2 for root console
As desired change root console is not available on tty2 anymore. For
backward compatibility gnome is still running on tty7. So, it's actually
running on two terminal. Consequently we cannot use tty2 as root-console
on sle 15. We add softfailure before bug is resolved and gnome is
running on tty2.

See bsc#1054782.