Add settings to set ADDONURL settings properly
For sle 12 sp3 we have SDK addon which doesn't exist for SLE 15. To have
single test suite we want to define these settings accordingly in
main.pm. In long term for SLE we may ignore addon and addonurl settings
as this part will be done by SCC and only product has to be selected
properly.