Improve extra_test_filesystem test suite stability
In snapper_cleanup test we have an issue that sometimes serial output is
clogged with some irrelevant messages. See run
https://openqa.opensuse.org/tests/470908#step/snapper_cleanup/187
And then we try to compare digit with the string which gives irrelevant
error. The easiest solution will be to just grab biggest number from the
output, as space is printed in bytes.