Adjust external iso test for storage ng
Basically this test requires selection of first drive, what was the case
in choose_disk module, which is replaced by firstdisk module as it is
already adjusted to storage ng. Another modification required is that
there is select disk screen which appears when any disk is not
formatted.

See [poo#23340](https://progress.opensuse.org/issues/23340)