Fix staging and adjust tests for new welcome screen
For staging we don't want to add all modules, as we do in
openQA Functional. This change allows to define short names of
modules to be added. E.g. ALL_MODULES=base,serverapp
In case of value '1', behavior remains same.

Another change for staging contains product selection on welcome screen
and not on separate screen as it was before. It's relevant for staging Y
only at the moment, but we can enable it for production once change is there.