Remove workaround for dependency issues and missing patterns
Now issues with dependencies should not appear anymore and hence we
should not soft-fail them. Point here is that original bug was created
for leanOS. If we test SLES, we should create relevant bug. To resolve
original issue, we add modules which are expected to be in installation.

See [poo#23938](https://progress.opensuse.org/issues/23938).