Handle changes in evolution 3.26
Prior to Evolution 3.26, the 'new account' assistant was launched
unconditionally. Since 3.26, it only starts when the 'Mail' register
is being accessed (which is the default register selected when
starting evolution without paramteres, like for example -c calendar)

So, newly, the main window starts up, the Mail register is auto-selected
and then the 'new account' assistant launches.

This requires now that, once we cancel the new-account assistant, that
we have to properly handle to close the Evolution main window too.