Work bsc#1058099 around to register against proxy SCC
Problem is that before we select product, it's considered to be leanOS,
which we cannot register against SCC. With regurl, seems that we are
trying to do so. Which leads to an error, which we don't expect to be
resolved soon. This issue replicates on multiple architectures.
With following parameters:
SCC_ADDONS=base,script,desktop,serverapp SCC_REGISTER=installation
we can get rid of step when add repositories manually on addon products
page.

See [poo#23492](https://progress.opensuse.org/issues/23492).