Adapt "default_desktop" for SLE15 after SCC can be used
With recent product changes the default desktop of SLE15 depends on the
selection of modules during installation. The availability of modules depends
on registration. The product selection of "leanos" is not a valid choice
anymore so we can safely remove that as well.

Verification run: http://lord.arch.suse.de/t7276

Related progress issue: https://progress.opensuse.org/issues/25282