Make yast2_ui test flow easier to review by splitting into two flows
The test scenario "yast2_ui" is infamous for showing a lot of valid product
bugs but because the yast modules are used for configuring so many different
product services the test is prone to fail often and therefore hard to review.
By splitting into the ncurses and the GUI tests the review should be made a
bit easier.

This commit replaces the test control variable "Y2UITEST" with two variables:

* "Y2UITEST_NCURSES": Call ncurses based yast test modules
* "Y2UITEST_GUI": Call GUI based yast test modules using Qt/GTK framework

Verification runs:
* yast2 ncurses: http://lord.arch/tests/7499
* yast2 gui: http://lord.arch/tests/7498

Related progress issue: https://progress.opensuse.org/issues/19922