Do not run 'register_and_check' for controller
The controller node is not part of the CaaSP cluster but it is
used as a helpful mechanism, acting as dhcp, dns and ntp server.
It also operates on SLES 12 SP2, which is irrelevant to MicroOS.
As result, the SCC registration against CaaSP fails for obvious
reasons and brings the whole testsuite into pieces.

This PR removes the 'register_and_check' from the controller
and also fixes the `CaaSP DVD Deployment` when the `REGISTER`
variable is set to `installation` string value.