Readd get_x11_console_tty supporting all current variants (#3645)
I moved the distribution initialization to make sure the variables evaluated
in init_consoles calling get_root_console_tty already regard the updated
defaults, for example the value for default desktop.

This does not consider a display manager different then the default for a
certain desktop as well as other potential impacts but the approach is proven
to work for now.

* SLE15 gnome: http://lord.arch/tests/7577
* SLE15 icewm: http://lord.arch/tests/7564
* Leap 42.3 QAM icewm: http://lord.arch/tests/7562
* Tumbleweed gnome: http://lord.arch/tests/7563