Fix user#01 selection on gnome
On gnome when having many users created, not the first user in the list
is selected, as mouse cursor is in the middle of the screen. Clicking
doesn't work, as list gets scrolled when pointer is moved, as a result
we click wrong user, so using up key. Whereas, to select user correctly
needle has 100% match level, as matches other users with 97% accuracy.

See [poo#25446](https://progress.opensuse.org/issues/25446).