x11_start_program: Speedup with optional check for target match
Adding two options 'target_match' and 'match_timeout' which can be used to
check for a target needle with assert_screen instead of just waiting for the
desktop runner to disappear which is a waste of time especially when after
x11_start_program finishes and the next test call is an assert_screen. This
can now be done within x11_start_program to save time.