Stabilize yast language module test
On o3 test module fails sporadically as yast gui doesn't show up in
time frame of 60 seconds. Afterwards we also not able to collect logs as
rpm verify fails to execute in time frame of 300 seconds which blocks us
from investigating root cause of the issue.
On top of that we've added code to kill all running yast processes which
may be there from previous test modules and may affect results of the
test.

See details in [poo#26104](https://progress.opensuse.org/issues/26104).