Test libcamgm via yast2-ca-management in x11
This case is based on below test case:
https://bugzilla.suse.com/tr_show_case.cgi?case_id=1595188
and it can be run successful in non-fips mode, failed in fips mode with exist bug:
https://bugzilla.suse.com/show_bug.cgi?id=1033977
non-fips veriy run:http://10.67.18.165/tests/161
fips failed run:http://10.67.18.165/tests/160

submit this automated case in fips mode so that can push solve exist issue.
If somebody use it in non-fips mode, it can be also added in.