Fix prepboot fs type selection
The initial problem is that we go through the list from default postiion
and not from the top or either bottom, so prep boot fs type could not be
selected. Another issue is with fs types which miss id and installer
crashes, to work this issue around we go from the bottom of the list and
not from the top.

See [poo#25460](https://progress.opensuse.org/issues/25460).