Fix s390x jobs stuck in shutdown until openQA job times out
Whenever the SUT is expected to disappear we need to ensure the remote
connections are terminated in before to prevent the test to be stuck.

With this change also renaming `prepare_system_reboot` to
`prepare_system_shutdown` as that is a more generic term (reboot = shutdown +
boot). Also it helps with abstraction to call `prepare_system_shutdown`
unconditionally as the method takes care internally to call appropriate
actions which may be backend specific. For the normal qemu case the function
is a no-op anyway.

Verification run: http://lord.arch/tests/7905

Related progress issue: https://progress.opensuse.org/issues/26886