disable_grub_timeout: Make sure the overview page is reached again at end
This should fix the many failures we see in 'start_install' which immediately
presses 'alt-i' as soon as the test module starts. As 'disable_grub_timeout'
is the one immediately before and it just presses 'alt-o' to exit the
bootloader configuration but does not check for the next screen the subsequent
'alt-i' most likely gets lost.

Related progress issue: https://progress.opensuse.org/issues/26866