Rely less on serial console and snapper's CLI API
On Hyper-V we have inherently unreliable serial console, limiting the
amount of data we send to it makes tests on Hyper-V reliable.

As `snapper` does not have a snapshot query tool we have to use CLI
"API" but we perhaps don't have to rely that much on it's current
format.

Fails here:
https://openqa.suse.de/tests/1242169#step/installation_snapshots/3

Validation run on SLE15:
http://assam.suse.cz/tests/754#step/installation_snapshots/2

I'd like to validate in other situations but JeOS is broken wrt
snapshots and other test suites I was not able to find in OSD.