Adjust autoyast profiles for SLE15
For SLE15 we have another design for installation and consequently non
of existing profiles will work as before. As simplest solution we add
all packages medium dvd and add 3 addons as before. As an ultimate
solution we need to provide registration details, but this is blocked by
changes to the framework

See [poo#23736](https://progress.opensuse.org/issues/23736) for more
details.