firefox_audio: Go back to calling firefox with direct URL to prevent mistyping
This test module is about testing audio playback and should not try to do
anything else. With the recent rewrite we introduced a slow down, an increase
of complexity and also the problem that by typing into the address bar a
complex URL characters can go missing. Most likely the reason for that is that
firefox makes suggestions after the first character is typed so in many cases
the second character, which is 't' of 'http' has gone missing.

Verification run: http://lord.arch/tests/8035#step/firefox_audio/3

Related progress issue: https://progress.opensuse.org/issues/25654