svirt: Boot from CDROM or HDD where apropriate
Previously on svirt we booted first from HDD, which either failed in
case of installation from CD, and fallbacked to CD, or passed in case
of preinstalled HDD. This was a hack. Now we try to be more aligned with
QEMU's 'boot once' option by defaulting to CDROM where we have ISO
defined and after installation we change the default to HDD, or we
default to boot from HDD where HDD_1 is defined. (Hyper-V still
sticks to the old aproach, because it's a bit harder to fix it there.)

Validation runs:
* SLES Xen PV: http://assam.suse.cz/tests/818
* SLES Xen HVM: http://assam.suse.cz/tests/810
* JeOS Xen PV: http://assam.suse.cz/tests/820
* CaaSP Xen PV: http://assam.suse.cz/tests/822