Harmonize the main.pm files further to DRY
* DRY on openSUSE/SLE load_ssh_key_import_tests
* DRY on openSUSE/SLE load_inst_tests
* DRY on openSUSE/SLE load_console_tests
* DRY on openSUSE/SLE load_x11_tests
* DRY on openSUSE/SLE x11regression tests

The impact of the change should be only on the schedule so I mainly compared
the evaluated schedule rather than look into the results from the test
modules. Mainly the change is a refactoring with no visible impact on actual
test execution. However, in some cases the order of modules execution was
changed for either SLE or openSUSE to harmonize both.

Verification runs:
* openSUSE staging cryptlvm: http://lord.arch/tests/596
* openSUSE TW textmode: http://lord.arch/tests/601
* openSUSE TW textmode: http://lord.arch/tests/591
* openSUSE TW gnome: http://lord.arch/tests/595
* openSUSE TW kde: http://lord.arch/tests/603
* openSUSE TW lxde: http://lord.arch/tests/593
* openSUSE TW lxde: http://lord.arch/tests/598
* openSUSE Leap 42.3 maintenance gnome: http://lord.arch/tests/589
* SLE staging gnome: http://lord.arch/tests/597

Related progress issue: https://progress.opensuse.org/issues/15132