Add module which partially dumps the proc and sys filesystems
The /proc and /sys directories contain lots of useful information which is
only available on a running system or a crash dump. In order to get this
information, we would usually need a system image or an actual live system.

However a system image or the actual system may not be available when it comes
to debugging so instead we can attempt to dump the contents of these
files.

Many of the files in proc and sys are not meant to be read from or
block until an event occurs. I am not aware of any programmatic way to find
out which files are readable in a useful way, so the script attempts to read
all of them (expect process files) and gives up after a timeout or if the file
produces too much information.

In order to speed up the process a whitelist can be generated which contains a
list of successful files from the previous run.

This module uses a shell script I have created for the LTP.