Add a workaround case for SLED to fix poo#28240
We are going to switch the default GNOME session to GNOME Shell after
Beta3 is released. But at the same time we are also migrating test
cases under x11regression to SLED15， however these cases & needles
could be affected by the unimplemented feature.

It's better to add a workaround now which will switch the default
desktop session to GNOME Shell after the installation of SLED.
This will help us migrate the x11regression cases for just one time.

  see also: poo#28240, fate#324384