LTP: Make test execution more robust and better results
Better integrate the LTP with os-autoinst so that individual LTP test cases
are treated more like native OpenQA test modules. This allows milestones to be
used effectively, so that if a test crashes the SUT, the proceeding tests will
still be executed after restoring a snapshot.

It also improves the appearance of LTP test sets and makes the progress bars
far more accurate.

Also this updates the JSON result array format to include an overall
environment field which dramatically reduces repetition in the result files.