zypper lifecycle: Hard-code package on JeOS
For JeOS build testing we are always using the latest repositories,
because JeOS images are build with a *different* build number to
SLES/Leap. It seems that SLES repositories are not populated with
packages for several hours (I have seen the test pass after 14 hours
from the image creation), and actually no package in the image comes
from any repository - it's from the image. So we hard-code
'sles-release' package, and it... works. Somehow.

Fails w/o the workaround http://assam.suse.cz/tests/1289, passes with
the workaround http://assam.suse.cz/tests/1291 *at the same time*.