Increase the timeout
The QAM updates that contain the new version of the docker images
are quite big - nearly ~10G. As a result it takes more time to
complete the transactional update.

This PR fixes poo#28633 by increasing the timeout up to 20 minutes
for the transactional-update and also 5 minutes per node for
Velum.