Improve overall stability of zdup tests by recording less screenshots
Looking for 'Retrieving' without the colon will record at worst two
screenshots per installed package, 'Installing:' will record at worst another
one per package. This yields for a normal "Leap to TW" scenario around 1200
screenshots, e.g. as seen in
https://openqa.opensuse.org/tests/549274#step/zdup/1198 .
Saving these jobs, loading and displaying them is putting a strain on the web
UI without a lot of value because it is unlikely human reviewers want to find
the right screenshot from 1200.
Trying to reproduce these scenarios on machines which have less performance
then can be even near to impossible ending up in hard to explain "incompletes"
as observed on lord.arch.

This problem is adressed by only looking for 'Retrieving:' so a potential
reduce of screenshots during the retrieval phase of factor 2 as well as not
looking for individual installation results reducing the screenshot number
from the aforementioned 1200 to 350 while being able to actually finish this
scenario for the first time on my machine.

As an individual timeframe in a video can be selected and referenced now more
easily in openQA this should not be a problem.

Verification run: http://lord.arch/tests/8174

Related progress issue: https://progress.opensuse.org/issues/25894