LTP: Replaced piped tee from zypper_call() by dumb_term parameter
+ add dumb_term parameter to the rest of zypper_call() where '| tee'
wasn't used.

Adding '| tee' to zypper_call() was inspired by commit
2e114fed ("Move iputils package from setup network to install LTP")
which added pipe to script_run().

But adding it to zypper_call() is wrong because it does not fail on
error due pipe. This is problematic as we use zypper_call() to install
packages without the testing does not make sense (e.g. git for cloning LTP).
Tests without these packages usually fail anyway, but later which is
misleading.

NOTE about boo#1055315 ("Zypper outputs ANSI characters regardless of
termcap") mentioned in 2e114fed: keep the workaround as the fix is
available only for Tumbleweed.