Simplify the loop in ensure_installed
pkcon is run in a loop to avoid temporary glitches to become problems,
but the needle to check the prompt for confirmation was only checked
once. So the loop was fruitless.

So I change the pkcon call not to prompt and make the output more
readable by adding -yp to it