Disable online repos and add o3 mirror for openSUSE
For openSUSE it's hard to replace repos which we use. In most of the
scenarios it works fine as we don't install packages which are not
available on DVD or mirror in case of net install. Whereas it causes
issues in some scenarios when we install additional software in the
tests.
There are 2 parts for the solution, one is to modify repositories in the
installer, which is possible only by tweaking control.xml before we
trigger installation, or by at least disabling online repos which point
to download.opensuse.org. It's easier to handle scenario with installed
system and we already have test modules which do this task. As of now we
enable this behavior with feature variable and for test suites which
fail due to this particular issue. After we see that it works fine, we
can apply it as default behavior and exclude some scenarios if needed.

See [poo#17436](https://progress.opensuse.org/issues/17436).